requirements.txt
# Django main
Django==1.3.1

# Postgresql
psycopg2>=2.4.1

# Test coverage
coverage>=3.5b1
django-coverage>=1.2

# I18n and L10n libs
transhette>=0.5.10
django-modeltranslation==0.3.2
